# news-aggregator/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a list of articles",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search articles by title or content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Technology"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter articles by published date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-06-30"
                        }
                    },
                    {
                        "name": "category_name",
                        "in": "query",
                        "description": "Filter articles by category name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Business"
                        }
                    },
                    {
                        "name": "source_name",
                        "in": "query",
                        "description": "Filter articles by source name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "BBC News"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Articles fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "category": {
                                                                "type": "string",
                                                                "example": "Technology"
                                                            },
                                                            "source": {
                                                                "type": "string",
                                                                "example": "CNN"
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "AI Innovations"
                                                            },
                                                            "content": {
                                                                "type": "string",
                                                                "example": "Latest trends in AI..."
                                                            },
                                                            "image_url": {
                                                                "type": "string",
                                                                "example": "https://example.com/image.jpg"
                                                            },
                                                            "published_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-06-30 12:00:00"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 500
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category or source not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category not found"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while fetching articles."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a single article by ID",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "source": {
                                                    "type": "string",
                                                    "example": "BBC News"
                                                },
                                                "aurhor_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "Technology"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "AI Innovations"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Latest trends in AI..."
                                                },
                                                "image_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/image.jpg"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-30 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while fetching the article."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Enter your API Token to access protected routes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}